Add File.new command.
Lists can be accessed now with the same way as arrays, using indices
enclosed in square brackets.

Allow zero length arrays.

Now command scripts can query and use the standard input.

Despite the future development of the language, the desired direction
is towards a free form of coding. This commit introduces and use the native
append function to implement such expression as a proof of concept:

    append "x" in m as "k"     # append a string to a map with k as the key
    append "zxc" in a          # append a string to a string array
    append 20 in i             # append to a number
    append "asd" in s          # append a string to a string
    append 2 in l              # append to a list
    append "qwe" in "zxc" as v # concatanate two strings and associate the result with a symbol