This is to save development.
As usual the diff is quite big, so the log is too long to fit in this commit header,
plus it can not track all the details.

However, this is mostly shell development, both to its readline interface but also to
its semantics that changed quite a bit.

Still there is no documentation, since the dust has to settle to make up decisions.

Numerous other development bits to other than shell related units.

And there is also a cplusplus wrapper sudoku solver and generator:
https://github.com/vaithak/Sudoku-Generator
This is really C with classes! Thanks.