Add a minimal snprintf implementation.
However this is just the first bit towards an alpha namespace.
Not sure yet for the semantics, so this is just the first little stone in the water.

This implementation is from:
https://github.com/tatetian/mini-snprintf

Very nice simple to understand code. Its LICENSE though hasn't been provided. The
attribution to the author contribution to the code evolution is a big humble thanks.

It is supposed to be the min[imal] snprintf() implementation of this system,
and certainly is with almost 200 cloc!, so it supports just %s, %d, %ld, %lu, %%.