This might fixes the long standing bug in the editor, which no matter what I couldn't reproduce with a reliable way.
It seems that the cause of this was that the "video_first_row" property of the structure,
sometimes ended up with a reference to a free'd row from a "delete line" operation
and when the cursor was at the first visible (video) line on the screen.

The real fix, was waysss too complicated because the algorithm is quite weak,
and it had to be applied to too many conditions of the code to be short and
correct. As even then and because opossite operations (insert|delete line)
could conflict and easy confuse the algorithm, there wasn't a strong guarrantee
that it will do the right thing. So instead of this overcomplication, we just
choosed to set the pointer at one place, and when we set the "video structure"
with the proper buffer lines, just before the final drawing.

Quite probably this suffices, but time will tell, that finally this was the
underlying reason for the crash in the undo/redo operations.

The fix started as a easy convenient workaround, until i realized that this is
actually an optimization, because it might be as well the case and a suspition,
that this property it might not be needed anyway or anymore (it was there from
the very first days of development and when i was trying to handle some obvious
conditions that a visual editor is facing and needs to handle, and this  was a
quick and easy way to do, not the proper way!). So if it works there isn't such
a big will to optimize it, as i do not want to touch this code never again :).