L: parse escape sequences when parsing string formats.
Make 'include' a synonymous for 'loadfile', as for a clear distinction
with 'evalfile'.

Do not include changes in the documentation and do not add test, neither remove
the IO dependency which technically is not required. This is not for development
reasons, it is just a part of the process in the middle of the transition to
the next development branch.

Add string_to[upper|lower](). Those should be included with docs and tests though.
But there is no time!