Handle dependencies when making a target.
To do that we had to overcome weakness of the recursive algorithm (if you could
call this thing an algorithm, as it is nothing but a hack - the whole thing is a
hack anyway).

After a bit of time, i fixed two different cases, to let it pass segfaults, and
at the end it left with just memory leaks, but it completed the task and those fixes
seems that they don't influence the already written code. At the very end, i used
a strange way with external objects, which makes sence actually, to avoid the
memory leaks, by using the function nesting mechanism, which is a nice way to
encapsulate and hide the details while working with data, which it is going to
be freed when exiting the outer function.

This info should be included in the documentation too. And as a lesson on how not
to use a programming language, that it has written by someone without a proper
education, that started programming after forties++ and C at 52, that he never
wrote such thing before, that still hasn't realize fully the programming concepts,
and!!! with automatic resource managment.

Fortunately the written code is already too complicated, so it probably can not be
much more complicated than this, and which and at this state of time, still does
(probably) the exact thing we ask from it, without segfaults and memory leaks.

Albeit with deep holes in the underlying logic and fragile balance, exposed brutally
by this hack that just happened to work, with this simple and expressive way, and
which we love a bit, and which we could do it even more expressive, if it could
make a sense in time, which obviously it doesn't.