The first bits of the migration to the new namespace.
Add a 'PACKAGE' target to Make utility, to use the packman library to build an external
package.

Add a verbose facility (a callback function) to our tiny fetched library.

Add a bootstrap Packman target to download and build libcurl, without the need of an
external tool.

Add the ability to "loadfile/include/import", to load symbols to the first found outer
function, from an if or loop block.

Try to print as much as possible unended blocks, to give a better hint.