Enhance the foreach loop for lists, to match the semantics of arrays, so that can work with indices and values also if desired.
Get the index from list members of a map.

Relax the do/while loop condition requirenment to be inclosed within parens. Those
can be end on a NL or with a semicolon. If the condition spans into multiply lines
then this is the case for parens, otherwise the week parser can not find the end
of the expression. That was the last case where parens were needed for evaluations.
Now the free form without semicolons and parens for statements and expression, it
seems that it has been stabilized enough to be considered as established semantic.

Enhance the Make utility with a function that can compile and assemble only objects
and link them at the end of the build all together. This is crucial to achieve what
we want from now and on. To cut development time devoted to a compiling cycle (when
we are developing new code or when we (and quite( usually want ot look for the fbug
of the old code), and to split the capabilities of the 'z' namespace into functions
(where a function it is and should be considered as the minimum code to achieve the
requested capability).