Supply the language document with the code that is intented to correspond with its semantics.
For now it is the copy of the tests. Some of them may fail because of some assumptions
about the top directory. Still they have to be written again with the established and
stabilized development, so they do not really correspond to the declared intentions of
the document, but still it is something for now at least, though not that representative.