Replace the `loop` loop, for the `for nth times` idiom.
Simplify this loop, the `while` loop and the forever loop by rejecting any other statements
during the intialization.

Do not require parens for the `while` expression.

Enhance the list members of a map, to set a value using indices.

Fix a bug when setting a function variable through a function call, using a map
member.

Develop a bit the new semantics tests to match our intentions.