Do not require the bars when declaring the symbols in `for ... in` kind of loops.
We now require parens, only in the C's for loop or only if the evaluated expression
in do/while loop doesn't fit in a single line.
With this and with the last development, syntactically we are now somehow close to
what we had in mind about the core part of the language.

As a sife effect, this should speed up many operations, but this was never a priority
and now its too late to ever be. If we want this, we should rewrite the code from
the scratch, which of course it will not happen anytime soon. Unless we'll turn our
direction on the wide ocean highway drive, which is obvious by now that won't happen in
our times.