Use {[...]} when creating inline lists, as supplementary to list ().
This is not that perfect, because of absence of symbols that make sence,
are not used elsewhere and are standardised.

S-Lang uses '{}' and from what i saw java uses a pair of braces '{{}}' for
arraylists.
We can not use '{}' because we use them to declare maps, and we could do it
like java does, but we opted for '{[]}', just because it seems that it makes
a bit more sense, as a list is like an array, except that is a heterogenous
container.

'[[]]' or '[{}]' or '([])' or '[()]' or '<>' or '<[]>' have conflicts too with
other constructors, except the last that could afford a tiny bit of overhead to
check, though noone else uses this too.