Implement arrays of lists, which is the last major functionality that left unimplemented.
All the old code and the tests should all pass, but might there are still unhandled
cases. If so, we will try to fix them with time, if time give us time, but only
time will tell anyway.

It looks that and with regards to the syntax and semantics, from now and on it is
mostly salt and sugar, but only if there will be a need for this extra flavor,
which is unlikely, unless we really like to do it, even if it is about cosmetics,
or if it is just for expressional reasons of which is not unlikely.