Add a syncdir module, that internally uses File.copy with update, force, preserve and recursive options.
This best should be exposed as a library, though it still is C.

This syncs only directories. We need also to expand it with and probably exclude_files
(currently it can exclude directories) or|and with an extra function reference
to check for older, ... etc, and with a hashing as an extra check, but especially
with interactive options, for the full functionality and convenience we know from
our old S-Lang script (we need this for at least when removing extra files and
directories).

This can be used as:

import ("syncdir")

Sync.dir (src, dest; verbose : 1, exclude_dirs : [array of directories])

Exposed as a script called as:

Syncdir src dest -v --exclude-dir=adir --exclude-dir=bdir

As a sample we can use a script to sync ourselves in a usb stick:

Syncdir z /usbstick/z --exclude-dir=sys/tmp --exclude-dir=sys/i686 --exclude-dir=sys/x86_64

It is probably a good idea for a Sync.me. This to work best we have to put all those
cached.build files to a database, so they shouldn't be syncronized, as those are
specific to their machine.