Add an archive module, which at time should be exposed as a library.
For now it can extract gzip compressed tar archives.

It is based on libdeflate https://github.com/ebiggers/libdeflate
and microtar https://github.com/rxi/microtar.

Many thanks.
As a general note, we favor and promote small codebases even if they are unknown
or experimental. Usually in the process to implementation, we may understand a
a bit the algorithms. That was the case especially for microtar, that we had to
research for the tar format specification.