This hopefully should fix a rare and hard to reproduce bug from the very begining, in interactive sessions to the `substitute` editor command.
This was somehow a bit of strange, just because there wasn't a way to print some
values for debugging, as in that case the bug wasn't triggered. Still there is
no idea for this behavior.

Looking at this code from the very first C days, of course it could be written more
efficiently. Generraly the codebase now, is just too enormous and complicated to
handled by a single human old fool full of illusions, that even the challenge doesn't
trigger some kind of satisfaction. This might looks like karma yoga, though even
the mention and inner thought procedure, can fade away the practicing, though it
was really an unconscious state of exersising life. Everything looks too strange
anyway, though there is some kind of logic, only that is too crazy and undescribable.