Add an oversimplified find module.
Use this to expose a Ls command for now.

The module can use much more optimizations, as every little code in this codebase
in fact, but from a quick test just for the curiosity, its execution speed it is
roughly the twice with ls coreutils, so it looks perfect considering the fact
that is few hours of coding, when its counterpart is under decades of development.
Not that it matters much though, but generally and especially the main tools like
the editor, they really use very gently the resources, and this it was really a
a matter that matters, from the very first scratch code.