Extend Argparse.add () to accept a "defval" qualifier.
This show the power of the qualifiers, as you can extend an interface without
changing the API, as you just have to handle the qualifier. Again the attributes
for this mechanism, belong to John in SLang, where we used to use it much.

Use this to handle --uid= and --gid=, where we want either -1 or > 0, as in other
case the defalut value zero could mean the root user.

So, we used this to extend the find-module to match [ug]id. Additionaly we can match
only executables.

Now we handle even more gently the memory sources into that same module, by storing
a pointer that points to an already allocated string. Little bit more complication
but it worths, as again the first and probably the only one, priority is exactly
this, and i think we're doing well here. Our editor consumes ridiculously really
few resources and yet is quite efficient even for today standards. When our view
on this - the whole idea of this project anyway - is with 1980' terms. So probably
(not probably for almost sure) it would be the most advanced of the time, especially
if you think that we use own code (to handle even the terminal) for every little
everything, so totally indepented (other than libc ofcourse). Even now there is a
posibility that we could do it, even without libc (if the time is generous maybe we
can do it one day, as the code is here).

Also added a Find command, but which is (almost) exact Ls, as we yet have to decide
how we can perform a function on the returned array, so this just to introduce it
in the repository and develop later.

As last, here is a personal script written in our language, which is being used to
sync automatically a usb stick. The expression (syntax wise and semantics, is quite
close to what we wanted from a scripted language. There are still a couple of things
that can be extended, and a couple new additions that will add a lot of value in the
language, but this should wait to the next iteration (if ever comes).

import ("path")
import ("syncdir")
loadfile ("argparse")

const concat_with = path_concat

var me = __argv[0]: path_basename ()

var argparse = New Argparse (3, 0, me + " [option[s]] source-directory destination-directory")
argparse.add ("to_usb", 0, "to-usb", "sync to usb", BooleanType, 0)
argparse.add ("from_usb", 0, "from-usb", "sync from usb", BooleanType, 0)
argparse.add ("help", 'h', "help", "show this message", BooleanType, 0)

ifnot ok is argparse.process (__argv, 1) then exit (1)

if argparse.exit is true then exit (0)

ifnot argparse.results.to_usb + argparse.results.from_usb {
  argparse.usage ()
  exit (1)
}

const hom   = "/home/user/dir"
const usb   = "/media/removable/user/dir"
const src   = if argparse.results.to_usb then hom orelse usb
const dest  = if argparse.results.to_usb then usb orelse hom

const exclude_dirs = ["ex1", "ex2", "bla"]

func sync (v)
  return Sync.dir (src: concat_with (v), dest: concat_with (v); verbose : 1, interactive : 1, exclude_dirs : exclude_dirs)

const arr = ["d1", "d2", "d3"]

for v in arr
  if sync (v) isnot ok then exit 1

I think it's not possible to be even simpler than this, or more readable. As what
it really counts when you met a new code, is to realize the intentions. Though the
interface and the other background code is not that simple, as it looks!

Dedicated to the human beings that suffer and pain, facing the inevitable. They and
we, have to know that there is more that this we understand right now. Also to those
who died laughing when they knew that they are faced with the unjustified and they
died hoped that they will touch the world with such way, so the world will realize
it one day (so they died blessed to see that day coming). And so if really there is
this "Sun Of Justice", then this day will come, and probably will come soon.