Add to Make utility the ability to look up for local spec. Add the --local-spec switch.
This is to provide the same functionality, for local specifications, both for testing
reasons, but also to add personal functions.

Document a prossible bug that is not reliable reproducible yet, and when reading from
standard input, as the code is probably assuming too much. This was exposed using the
Search.string command. This is because there is some lack of fundamental knowledge of
low level functions, i believe. Unfortunatelly this is probably isnot going to change
anytime soon, though it seems its enough to keep going, but we are vulnerable anyway.