Relax the rules and allow empty value members, while setting a map. In this case they get a null value. Currently this was done only for new types and qualifiers. This might break a thing, though there is no serious reason.
Also, we couldn't use `const' for multiply assignments. This should fix it.

Not all of the binary operations that check for equality, and evaluated to false
and the next is an AND operation, were interrupted. Do it this now. Hopefully it
is that simple.

Now, all the types should have the chance to compare for equality with at least
null values.