Make: add pre/post build callback functions.
The implementation though it is not that perfect, as since there is
no reliable way to get the exact scope of the map, they have been
defined in the public scope. At least they have been declared as
anonymous, though it is not perfect either.

Add many library specs to zero namespace and build the editor library
as a test.