The implementation for map callback functions in 7ec3e2944c2142f99fc60ee399d6e811273f4b6d, allowed full access to the "this" instance, so we loose the encapsulation.
Fix this leak, by creating a dummy parent function that goes nowhere and without an
environment and then push the "this" instance to the stack. The only prerequisite is
that such callbacks should declare the map to the argument list as the first argument
and use that for access to the instance as usual.

Again with regards to the abovementioned commit message, this looks now an acceptable
solution, and probably it is also the way for any future closure implementation.

Also, you could actually define "this" as a var or as a const, so disallow this posibility.

Also, copy the argName array when copying a function.

Also, use "as callback" keywords instead for "as New" when setting a callback function,
which make more sense. This should break at least one library build but we'll fix this
shortly, if we blessed.