Finalize the grammar in Make by providing the desired consistency to all the members of the instance.
We did this by dividing the instance in (as many) submaps (was needed), so every little
everything is divided by a dot '.'.

We always begin with the instance name ("this"). Then the second keyword usually is
the function that we apply to the properties of the instance, and then the properties
properly divided by dots, separated by categories:

  this.make.shared.library ()

or if we just access the property without applying a function:

  for v in this.shared.library.sources

This is now an almost human way to express and probably quite acceptable by purists
and the academia universe. I do not think i belong to neither (for sure i'm not in
academia as i do not think there is a position for an old ex shepherd punk without
qualifications), i just want to find my way and follow the flow of my thought. Then!
i will resque this world as its our destiny, though probably there is no destination,
just an eternal evolution - if this world is not just an emulation and nothing really
moves on as we fakely believed, as we are (might) just doing eternal circles around,
and we believe that we are doing something. At the same time there is this ground that
smell like the earth does, and there is really nothing else like this smell and this
is real, or at the very least a very acceptable attractive illusion or reality (take
your pick or just choose your own confusion (and this is something that is not going
to change anytime soon (probably), as there will always be (probably) a doubt - i do
(probably) not)).

As a last, there is just one last big thing in Make that it isn't handled properly.
It is so crucial to fix as it is crucial to breath. After this, we are going to go
to the next iteration (the last one before the very last). But we are a bit far away
yet, yet so close. This world is full of conflicts, or for some, is just made of just
conflicts (antithesis in my native), that work like magnets, that might allow balance,
desired to keep us and our destination path, under control and some kind of form that
give us shape and sake to breath. Hopefully this is real and more hopefully we'll find
it soon. To be honest we better do it soon, so soon as the new moon.

It might worths to repeat it here loudly in the void, that this project's only ambition
is to expose intentions.