Add is_defined() function that returns true if the given symbols has been defined, or false otherwise.
Use this function to Make utility, so it can be used as a library. We didn't use directives
which is more natural, because the interface is weak and it might be removed.

The other Make changes are just cosmetics, because the image counts a lot!

But this is also to fix (hopefully) for good this time, the path_basename_sans_extname().
The mistake was so obvious, that i wonder if there are more like this one. I'm
pretty sure there are.