Fix the buf_set_row_idx() function from the editor library.
The following doesn't do what i thought it will do (the continue statement doesn't really continues the loop)
do {
  idx = DListSetCurrent (this, index);

  if (idx is EINDEX) {
    index--;
    continue;
  }

} while (0);

This caught by a script that restores a session. The cursor was on the
last line when the editor exited. Meanwhile the file has been shrinked and
when the session it was restored, the script called that function and naturally segfault'ed,
because the index hasn't been decremented as i thought it will do.

This was the first time that this block has been executed or the first time that
this condition has been met.

Actually there are two bugs in the old block, so it made me think that this was
just a bad time. Anyway continue and keep walking and stop talking.