Add string_to_hexstring() and hexstring_to_string() to libc.
This is to test libc but mainly for recording the convertions in the codebase.
The implementation is from https://github.com/rofl0r/libulz and they work only
in the ascci range, but still useful for educational purposes.

Add also a corresponded module.

This is also to fix array_sort () from std-module.

It also adds specific to this codebase errors when we generate errno, though
errno_string () doesn't handle them at all yet.