This integrates generatedLibc as part of the interface and exposes it as a script executable and as a library.
Use a post_exec() callback to add a call to the function when it is
installed (a bit of hack) via Make.
However it may be wise to be installed as a library and use it as usual.
Make doesn't handle that case.

Also just document, that it might worth to support something like:
load future from foo (using dots as separators (like in this case)).
And speaking for load also it can be used as:
var s = load as func filename
and combined as:
var s = load as func from foo bar
and possible:
var s = load as func from foo bar with (args, ...) though this requires
to change a bit the current parsing logic.

Also. End with the doubts, by naming libc as zlibc, and use the top source
directory as its default path in the hierarhy.
We should use in future specific libc profiles/specs, in any client
that uses the internal libc; then use a new Generator method to generate
the requested functionality, as a unit in the same directory.

La: fix the case when there is a dot in the file that is going to be
evaluated.

Make: start to use this.error (msg) for returning a failure with the
posibility to print the msg if not null and call errno_string() when
errno is not zero. It will be completed with time so there is a bit of
inconsistency between the expressions (it should be only one and only
one obvious way :)).

Use such functions everywhere in every unit. It gives clarity in the
code, plus makes better single statements, and so avoids braces (i do not
like them - if you can't avoid them for reasons, find nice ways to avoid
them, and this is one))