This is just for testing a retval interface. We use the time-module and one function that validates a tm structure.
This returns a map with a value and a msg fields. This adds a bit of an overhead but
it could be one way to do this kind of thing, if there is a desire or a reason, or if
we just want to be proud and heroes (and we might want to be if it is allowed by time).

Oh, and a bit of inline documentation as we promised some time ago.