Include stuff from cenv (the types macros mainly) in zenv.
We might not need all this stuff though.

For now we only typedef FILE without an actual stdio implementation, so we might
malfunction on errors.

Also the compiler will probably refuse to compile on static targets, if sees this
errno declaration, but didn't check.

But its ok for the sake of development, since we just use them for now in one target.