Modify the behavior of string_tokenize().
" a  b c   d  " should produce now ["a", "b", "c", "d"] with " " as separator

We do the parsing into the intrinsic function itself instead with Cstring.tokenize(),
because of a comment in that lib function, that didn't allow to develop the function in
in that direction (though i do not remember where it broke the logic).

This ended up as an optimization, as first it is minus one abstraction, plus fewer
allocations and releases. This remind me that there isn't any kind of optimization,
as it is mostly a functional prototype, than a mature product (quite probably we won't
have this chance); at the very least we could use own functions than from some generic
libraries (like the map functions, which they are being used heavily).

Anyway, further adjustments and refined behavior in this same function, it probably
could be done with qualifiers.

Also we might want to use an array for the token argument, as in such case simply it
could be considered as a set (still there is no such code in the codebase though that
need this functionality).