This is a first draft (buf functional) of a pager module.
This reuses existing libraries, so it can be optimized further with specialized
functions. It has also some fundamentals, to develop this unit with editing capabilities.

As a reminder, we are now a little past the end of our promised schedule (18 months),
so we are a bit late. As an excuse this is a complicated beast and with already
huge codebase (normally unhandled by a mere human, but we're trying). As another
excuse, we are missing an init mainly and some basic system files, so we are not
that far from the completion, and with more tools than we ever imagined at the
begining.

Perhaps this requires a new deep breath with a new defined scratch point, but it
is unclear if we want to walk back, or walk towards the happy path to develop
some visible functionality.