Abstract the pager module in a way that it could be easier to add editing commands if we are inclined to (though there isn't such a will right now).
It still needs some extra pointers into the buffer, to break the main loop into
small functions that could then be exposed, and alsp some ways to modify the
default behavior.

Also, it still needs to catch some signals, like the sigwinch but also others,
and return proper values and messages if the screen dimensions are too narrow
to be functional. Again, for now (and probably forever) we are good as we are
working with fullscreen terminals.

As last, attach a very minimal script to expose the functionality to the environment.
This needs to handle standard input and to duplicate the stdout fd, to handle
the terminal state as the process is missing a controling terminal.

Default keybindings:
 q    or Q          : quit
 gg   or HOME       : goto the first buffer row
 G    or END        : goto the last  buffer row
 k    or ARROW_UP   : goto the prev row
 j    or ARROW_DOWN : goto the next row
 h    or ARROW_LEFT : goto the prev col
 l    or ARROW_RIGHT: goto the next col
 C-b  or PAGE_UP    : goto the prev page
 C-f  or PAGE_DOWN  : goto the next page