Implement some of the todo items of the pager module.
Add also '0' and '$' motions to mean begin of line and end of line respectivelly.

This is now almost perfect, though there is no such thing like a perfect thing,
so yes there is a bug (or unhandled condition if you may prefer it as sounds gentlier).
Though a C program can be perfect (if you trust and freeze your compiler! at some
point (the last known perfect code interpretation)).

As this it can go very far, it might be a good idea to almost stop at this point.
This might means that this is the end, or it might signal the beginining (this
for sure sounds gentlier (strange as it sounds from an (almost) ex wild punk),
but happy anyway.

See you on the way.