But before we go, this is the very old good code from par:
https://bitbucket.org/amc-nicemice/par/src/master

This is an almost untouchable copy all of the original par code actual algorithm,
merged into a single unit and the functionality exposed as a library.

We use an instance and pass this to the functions as their first argument,
and we exposed this without hiding anything from the user (for flexibity).

However (for now) only the functions that expose functionality, changed their
signature and use this instance.

This is to keep as much of the original code with the exact way it was written,
so nothing has really been changed or removed (except the duplicated headers).

However the main function has been removed, but it is possible to use the (perhaps
known) exact same syntax par's command line interface, passed as an array of strings
to the exposed function (so to mimic and implemented the original standard behavior).

Usage:

  par_t *this = par_new ()
  par_parse_argv (this, array of strings)
  int par_main (this, string type input)

and

  var q = {
    just     : ifwewanttojust  # by default yes
    width    : ourwidth        # by default 78
    tabwidth : tabidth         # by default 8
  }

  var p = Par.new (; q)

  var argv = ["P=*-`_s#"]
  Par.parse_argv (p, argv)

  var s = some_string
  var r = Par.process (p, s; tostdout)  # without the tostdout qualifier this returns a string

also to see the corresponded options above issue from our shell:

  Par --help

and as an example, you may try:

  Par --just --arg=P=*-`_s#  README.md
    # you may need to escape some of the P option chars on another shell

  see par.doc for documantation and for their many many details!

As an inner deep expression which i would like to share , i feel that i've
been very lucky:

  - to read this very old code
  - to see it compiled by modern compilers in C11 without a single warning
    and turning on agrressive warnings
  - for the so many detailed comments
  - to read the documantation and laugh with the author about his apologies
    i'm totally sure that he should be very fine kind of a human being
  - and ... is very fast

so I feel to owe a way too much appreciation for this gift, many many thanks.