Use the pager code from par and unite functionality.
This is a rather incomplete commit, as while it is functional, we just use this
code to test some things first before we end up on the development route, which
it is also changing, begining from this commit, as if we continue the development,
it will look mostly like a scratch environment. As such the last commit should be
considered as the last stable.