Allow to set a map member through an anonymous function call.
While at it i noticed that this was legal:

var m = { a : 1 }
m.f = func { return this.a }

Do not allow the usage of "this" when calling methods outside of map declarations.

Before this had the additional property that you could get a private field through
a method call, thus it was missing the point. Nowhere in our codebase has being used
this idiom though.