Add autocompletion to the shell for the following cases:
   ./script
   ../directory/script
   ./directory/script
   /dir/script

This is triggered when the cursor is on the first token (word) on the command line,
and either the first char is '/' or ('.' and the next is either '.' or '/').

if `script' is a directory then it acts as autochdir, and trying to change the
current working directory.

There might be unhandled cases.