Add a very first functional completion UI that is being used when there are more than one item in the list.
This is a rather crude mechanism (the actual code is a stripped version from src/y/pkg/pager
adjusted for the shell), that displays one item per line.

It is not perfect (many small details could be improved) but it does our job for now, as
it makes our shell quite useful.

As this was one of the biggest missing features, this in normal days would signal a major
milestone, but it is not one of these days. All of sudden, many of our young guys are not
with their friends and parents as they should. We know next to nothing for death, but I've
a strong feeling that we'll know soon, and then the pain might soften a bit.