Use the tab key to continue with autocompletion, while in selection menu.
This will allow for much more faster navigation and less keystrokes (especially
with filenames). Also it might allow to remove a lot of code in future if it works
without flaws.

As this is really a hack, its usage might reveals bugs and unhandled conditions,
but the completion mechanism, it is a prerequisite for a shell to be useful these
modern days, so it worths the (even short) time it took till this point, or the
time will take during the waterfall kind of testing and development, but also and
the (rather few) lines of code it took for the implementation. Again you can not
say that it is a proper code!!!, it is really mostly a hack (albeit a working one).

There are still some details to fix or develop extra logic, though. But this is
for more sophisticated and advanced usage. We like simplicity as an excuse but
also as a belief. The expectation from a shell is to be fast and this is really
fast and with quite few resources yet (so we still have some more to spend if
we'll be blessed with time generosity).