This is to add extra logic to the completion mechanism as a first responce (to a previous git commit message), for many many details that need some handling.
  - set the exact pointer position for --long-arguments= that are awaiting
    an input (cut the extra space)

  - on the contrary, add an extra space for the initial command completion

We had also to modify once more the rline library, that is being used in our
shell and in interactive language sessions.

For those two features, they had to cooperate the zs.c unit with the said library,
with a special way, and so ideally we would like to be easy in future to spot and
understand, where and why, we did the change in logic.

The actual diff is small. But i add also a multiline comment and removed some
C++ style of comments in this commit. And i've removed a couple of uneeded pair
of braces in single line blocks too.

(Such commits really should be the very minimum, as in this case the environment
is fragile, as it is under development. And maybe it is a good idea to adopt this
wisdom at least from now and on (after so, so many commits with thousands of lines
of diffs (though it is quite natural when a system is under heave development))).

So i did it in purpose to emphasize this wisdom with an antithesis.

As last but not least, i've removed tabs also and i feeeel fine (i hate tabs with
a passion in code) and i can not understand how i didn't spot them after so much
time! Unexcusable!

* note that the actual diff, it is not that hard to understand anyway, but it is
  not easy either