zs: try to adjust better the cursor position, when it is in the middle of the command line.
Also fix some corner cases, like when there is leading whitespace during the completion
(which in that particular case, besides the wrong result, the application segfault'ed
because the length argument to String.append was negative).

Also various enhancements, like when the completion returns just one command that ends with
a '.' (dot). In that case the completion is triggered automatically (quite convenient
and fast).

Many more details need some care, but because the linenoise machine wasn't exactly
made for that much stretch, either we handle them with our own code to the underlying
machine (and we did add already a lot), or we end up with hacks. Either way is quite
fragile and hard to develop logic (like we do with our own readline). At some time
eventually and with practice, it will be stabilized, and then C will be hard to break.
And so we'll continue to invest in that readline implementation. In any case we've
been walked too far already in that path, with not perfect but with satisfactory results.