zs completion: handle ../[../*]char[s]tab, by prefixing the dirname to the returned result.
There is a chance that this will break other situations. In any case after this fix
there are not any known misbehavior in the completion machine (which of course is
a wrong assumption (only time will tell though (as it is too early yet))).

Fix gcc-10* two compiler complains, that they don't raised with gcc-12* and rightfully.

libe (normal mode): make the escape key to switch to the command line like ':' does.
It was long ago reserved for an extra mode (outer mode for various uses). If it
wasn't for that, this usage it makes perfect sense and is fast.