zs: fix a subtle bug while setting video rows and our string exceeds available columns.
This it could be observed when the cursor is on the last row with such string.

The fix was simple. Just set the first column.

Also catch early the condition and don't go beyond available columns, even if we
loose a tiny information, if the last character is a tab or occupies more than two
cells. We could do better, but doesn't really matter in this particular case.

Ls: handle multiply arguments.