La: set the semantics and implement +=,-=,/=,*=,%=,&=,^=,|= binary operations for arrays.
This works for string, integer and number types, but only integer arrays support them
all. For string arrays only the add operation has been implemented, and could accept a
string or an integer as their second operand. For number types only the first four of
those operators have been implemented.

Their first operands are the values of the array.